requirements.txt
fastapi
uvicorn
passlib[bcrypt]
pydantic
python-jose
pymongo
email-validator
python-dotenv
pyjwt
python-multipart
logging
